Service interface 삭제 및 ServiceImpl -> Service class 로 통합
